Move teams show to a folder and rename teams-form to teams manager Standardising purposes
